create a folder and separate returning and new user components
