+++ src/components/WelcomeModal/NewUser.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { Description } from "./styled";
+
+const NewUser = () => {
+  return (
+    <>
+      <h1>Welcome to Verse Clicker</h1>
+      <Description>
+        Click for verse, climb the leaderboard! Join the Verse community and
+        experience a world of endless clicking fun.
+      </Description>
+    </>
+  );
+};
+
+export default NewUser;
