+++ src/components/WelcomeModal/ReturningUser.tsx
@@ -1,0 +1,40 @@
+import React from "react";
+import { useTrackedState } from "../../context/store";
+import { formatNumber } from "../../helpers/formatNumber";
+import { formatSeconds } from "../../helpers/formatSeconds";
+
+import { Title } from "../Title";
+import Clock from "../Icons/Clock";
+import Star from "../Icons/Star";
+import { colors } from "../colors";
+import { H3 } from "../H3";
+
+import { DataWrapper, Stats, Value } from "./styled";
+
+const ReturningUser = () => {
+  const { returnData } = useTrackedState();
+
+  return (
+    <>
+      <h1>Welcome Back!</h1>
+      <DataWrapper>
+        <Stats>
+          <Title>You were away for</Title>
+          <Value>
+            <Clock />
+            {returnData && <H3>{formatSeconds(returnData.seconds)}</H3>}
+          </Value>
+        </Stats>
+        <Stats>
+          <Title>You earned </Title>
+          <Value>
+            <Star size={28} color={colors.shade80} />
+            <H3>{formatNumber(returnData?.cookies)} </H3>
+          </Value>
+        </Stats>
+      </DataWrapper>
+    </>
+  );
+};
+
+export default ReturningUser;
