+++ src/components/WelcomeModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./WelcomeModal";
