+++ src/views/Main.tsx
@@ -4,7 +4,7 @@ import Footer from "../components/Footer";
 import GameBoard from "../components/GameBoard/GameBoard";
 import Particles from "../components/Particles";
 
-import WelcomeModal from "../components/WelcomeModal";
+import WelcomeModal from "../components/WelcomeModal/WelcomeModal";
 
 import Layout from "../components/Layout";
 import Sidebar from "../components/Sidebar";
